pyproject.toml
[project]
name = "vib-data-process"
version = "0.1.0"
description = "FastAPI service for processing DOCX and Excel files with image extraction and OCR"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.23",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "openpyxl>=3.1.5",
    "openpyxl-image-loader>=1.0.5",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
    "json-repair",
]
